requirements.txt
sentence-transformers
numpy
pytorch
nltk
